Moved txToJSON to test/utils/helpers
