+++ typescript/test/deposit-sweep.test.ts
@@ -4,7 +4,6 @@ import {
   TransactionHash,
   UnspentTransactionOutput,
   Transaction,
-  txToJSON,
 } from "../src/bitcoin"
 import { testnetWalletAddress, testnetWalletPrivateKey } from "./data/deposit"
 import {
@@ -17,6 +16,7 @@ import {
 } from "./data/deposit-sweep"
 import { MockBitcoinClient } from "./utils/mock-bitcoin-client"
 import { MockBridge } from "./utils/mock-bridge"
+import { txToJSON } from "./utils/helpers"
 import * as chai from "chai"
 import chaiAsPromised from "chai-as-promised"
 chai.use(chaiAsPromised)
